Moving and refactoring rotate array extension to RangeReplacebleCollection
