+++ Sources/Extensions/SwiftStdlib/RangeReplaceableCollectionExtensions.swift
@@ -1,0 +1,50 @@
+//
+//  RangeReplaceableCollectionExtensions.swift
+//  SwifterSwift
+//
+//  Created by Luciano Almeida on 7/2/18.
+//  Copyright © 2018 SwifterSwift
+//
+
+// MARK: - Methods
+extension RangeReplaceableCollection {
+
+    /// SwifterSwift: Returns a new rotated array by the given places.
+    ///
+    ///     [1, 2, 3, 4].rotated(by: 1) -> [4,1,2,3]
+    ///     [1, 2, 3, 4].rotated(by: 3) -> [2,3,4,1]
+    ///     [1, 2, 3, 4].rotated(by: -1) -> [2,3,4,1]
+    ///
+    /// - Parameter places: Number of places that the array be rotated. If the value is positive the end becomes the start, if it negative it's that start becom the end.
+    /// - Returns: The new rotated array
+    public func rotated(by places: Int) -> Self {
+        //Inspired by: https://ruby-doc.org/core-2.2.0/Array.html#method-i-rotate
+        var copy = self
+        return copy.rotate(by: places)
+    }
+
+    /// SwifterSwift: Rotate the array by the given places.
+    ///
+    ///     [1, 2, 3, 4].rotate(by: 1) -> [4,1,2,3]
+    ///     [1, 2, 3, 4].rotate(by: 3) -> [2,3,4,1]
+    ///     [1, 2, 3, 4].rotated(by: -1) -> [2,3,4,1]
+    ///
+    /// - Parameter places: The number of places that the array should be rotated. If the value is positive the end becomes the start, if it negative it's that start become the end.
+    /// - Returns: self after rotating.
+    @discardableResult
+    public mutating func rotate(by places: Int) -> Self {
+        guard places != 0 && places < count else { return self }
+        if places > 0 {
+            let range = index(endIndex, offsetBy: -places)..<endIndex
+            let slice = self[range]
+            removeSubrange(range)
+            insert(contentsOf: slice, at: startIndex)
+        } else {
+            let range = startIndex..<index(startIndex, offsetBy: -places)
+            let slice = self[range]
+            removeSubrange(range)
+            append(contentsOf: slice)
+        }
+        return self
+    }
+}
