+++ Tests/SwiftStdlibTests/ArrayExtensionsTests.swift
@@ -94,26 +94,6 @@ final class ArrayExtensionsTests: XCTestCase {
         XCTAssertEqual(tuple.1, [1, 3, 5])
     }
 
-    func testRotated() {
-        let array: [Int] = [1, 2, 3, 4]
-        XCTAssertEqual(array.rotated(by: 0), [1, 2, 3, 4])
-        XCTAssertEqual(array.rotated(by: 4), [1, 2, 3, 4])
-        XCTAssertEqual(array.rotated(by: 1), [4, 1, 2, 3])
-        XCTAssertEqual(array.rotated(by: 3), [2, 3, 4, 1])
-        XCTAssertEqual(array.rotated(by: -1), [2, 3, 4, 1])
-        XCTAssertEqual(array.rotated(by: -3), [4, 1, 2, 3])
-    }
-
-    func testRotate() {
-        var array: [Int] = [1, 2, 3, 4]
-        array.rotate(by: 0)
-        XCTAssertEqual(array, [1, 2, 3, 4])
-        array.rotate(by: 2)
-        XCTAssertEqual(array, [3, 4, 1, 2])
-        array.rotate(by: -1)
-        XCTAssertEqual(array, [4, 1, 2, 3])
-    }
-
     func testShuffle() {
         var arr = ["a"]
         arr.shuffle()
